+++ core/src/com/pacman/model/ghosts/GhostRed.java
@@ -4,7 +4,7 @@ import com.pacman.model.World;
 
 public class GhostRed extends Ghost{
 
-	public GhostRed(float x, float y, World w) {
+	public GhostRed(int x, int y, World w) {
 		super(x, y, w);
 		setSpeed(1f);
 	}
